Moved the default CMake flags to the platfrom recipe files
This will help support different variants of Edge Core on different platforms.

COMPATILBE_MACHINE should be defined in the first line of file because it helps
Yocto know which file to parse when compiling for a specific platform. As stated here
https://www.yoctoproject.org/docs/1.7/ref-manual/ref-manual.html#var-COMPATIBLE_MACHINE

By default, UC_Hub firmware update is enabled on RPi and FOTA framework library on
iMX8 and Xilinux board.