During boot start edge-core only after network is connected and we have full access to the Internet
network-online.target - is a target that actively waits until the nework is "up". Usually it indicates a configured,
routable IP address of some kind. Its primary purpose is to actively delay activation of services until the network is set up.
Ref: https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

It is desired that as Edge Core is dependent on network, we should wait until we have IP address.

It is also desired that Edge Core should only start after Internet access. Added ExecStartPre to run nmcli
and check for connecitivity status. Only when status=full then launch edge-core.
Ref: https://developer.gnome.org/NetworkManager/stable/nmcli.html